Tighten the double click criteria.
If the distance between the two clicks are greater than 8
pixels it is not regarded as a double click.  This is in line with how
GTK+ detects double clicks.  There's an X root window property or
similar that defines the distance and ideally we should read that out,
but for now the patch brings the behavior a little closer to
what it should be.

(bzr r364.1.3)