Add CompNoMask to REAL_MOD_MASK.
REAL_MOD_MASK is used determine what bits of the modifier
state that should match for an action to be triggered. Virtual
modifiers without real modifiers assigned to them will use
CompNoMask as real modifier. Hence, the CompNoMask bit must
also be used to determine if the modifier state matches and
should be considered a real modifier.

Thanks to Andy Hefner.

(bzr r1309)