Don't use width/height <= 1 cursor images
These are (most often) bugged ones, and results in an invisible cursor.
This is an attempt to alliviate the XFixes bugs that causes certain
cursors to go poof. This is still a problem when NOT zoomed in. This
is a workaround that may or may not work. Feedback wanted.

(bzr r0.19.146)