Reduce animation cpu/gpu usage by damaging minimally.
At each animation step, damage only the parts of the screen that should
be updated instead of the whole screen. The smaller the animated window
and the larger the screen resolution, the higher the performance improvement
will be (especially tooltip and menu animations will benefit significantly).
This changeset affects all animation effects except Airplane.

(bzr r0.23.197)