Fix all possible freezes for good.
This adds a cycle check so that a cycle doesn't happen anymore during
focus animations, which was the source of all the freezes. Specifically
there was still a freeze that happened in some cases with gobby with
dodge set as the focus animation, and this is fixed now.

(bzr r0.23.272)