Improve performance with some animations.
This change mainly improves (on minimize) curved/horiz.fold, dream, glide, and
(on open/close/minimize) zoom and sidekick, by damaging a smaller region.
The damage region was being computed as the bounding box of the current and
the last drawing boxes. This was unnecessary as we just need the union of
those two boxes. This change implements that.

(bzr r0.23.304)