Comment out broken code for expo-handling
I wrote dontuse* because it is NOT the same as othergrabexist(). I'm
leaving this commented out because I should fix this to make eZoom behave
when expo is active. (This code disablex drawing of pointers when any
plugin EXCEPT expo was active, which is the opposite of what it's supposed
to do).

(bzr r0.19.218)