Force input focus on button 1/2/3 press.
Sometimes there is a case where the window is active on one screen already,
and because compiz is not multi-screen aware, when another window is focussed
on another screen, the window will continue to be "active" but not focussed.

Previously, when clicking on this window again, the window would not be
focussed again since it was already active.

This commit forces input focus regardless of the "active" state

(bzr r2622)