Don't unminimize transients when the window is being destroyed
Unminimizing the transients means that we have to loop the window list
in the ::~WorkaroundsWindow stage, and we don't know which of our windows
have already been destroyed (which is problematic for ::get ()). Since
the transients of windows are likely to be destroyed along with the main
window and and also during the entire plugin tear-down phase we are unminimizing
each individual window, there is no need to also unminimize transients for
each window.

FIXME: We should really implement a ::fini stage where we can guaruntee that
nothing has been destroyed

(bzr r0.392.106)