Don't auto-group windows if they aren't supposed to be a group window.
The auto-grouping code would use ::select, which ignores if the window is supposed
to be a group window because it was already done by ::checkWindow in the normal
selection path. This moves the checks into the ::select function.

The old behavior could cause strange apps that have two X windows but one "user window"
to behave strangely when tabbed - eg showing the tab bar but not actually allowing the user
to grab the tab, etc.

(bzr r0.169.410)