Don't increase leave pendingMaps at a non-zero number once the window is mapped.
We used to assume that for every MapRequest event that we got for which we also sent an XMapWindow that we would get a MapNotify.

According to the man XMapWindow this doesn't make any sense since trying calling XMapWindow on an already mapped window for which we have SubstructureRedirectMask set results in a NoOp anyways (so we'll only ever get one MapNotify and multiple MapRequests).

This fixes a case where no input focus would be set on closing windows with multiple remaining pendingMaps since we don't move the input focus if this is the case

(bzr r2646)