Don't stack mapping windows which requested no focus above the active window.
This adds a managing variable to PrivateWindow which represents the managing
state of the window. In addition, ::updateAttributes did not stack windows
below correctly if they were already directly on top of an active window
(since it is not valid to stack a window below itself, and the algorithm would
pass the below to the findValidStackSiblingBelow function as well as a
pointer to the active window (== below)). The function now compares the
below sibling to the window we want to be stacked.

(bzr r2609.7.17)