Make the switcher window being override redirect.
Previously the switcher window was a normal managed window (which
the switcher plugin actually lied about, see SwitchWindow::managed),
such that the decorators could pick it up and draw the switcher
decoration. However now with reparenting, this means that whenever
we actually manage and unmanage the switcher window (eg with map
and unmap) it means we must also reparent the window, which floods
the server with requests and makes the switcher slow.

Since we don't interact with the switcher window, it makes no sense
to manage it like this, so make it override redirect

(bzr r2609.7.44)