Set default value for vpSize before loading plugins.
There is a race where vpSize won't be updated to its "true value"
until we reach core's option change callback or we read the value
from the configuration plugin. Since we load all the plugins at once
on screenInitPlugins on the command line, some plugins require that this value
be set to *something* (It can be changed later) before they are initialized.

Fixes incorrect vertex settings with cube, causing it to zoom out infinitely
when switching viewports when loaded when compiz starts

(bzr r2609.7.77)