Release implicit grabs as soon as we handle events.
Implicit keyboard grabs which are created by the pressing and holding of a key
that was passively grabbed by XGrabKey are evil for a number of reasons.First
of all, there isn't any way to track them or to track which window they are
grabbed on so if a plugin makes an active grab and then releases it, the
passive grab is released on release, which means there is an inconsistency
between holding the key and having the grab. Secondly, if plugins actually
need a grab on all keys, they are better switching to an active grab
rather than relying on the passive one. Finally, if there is a passive grab by
another application by a key-modifier combo which has the same modifier mask
as the currently activated passive grab, then it means that this application
won't get its grab activated as long as the modifier is held down.

(bzr r2726)