Set total workArea based on the union of all workAreas rather than based on the position of struts.
Rationale: If there are struts positioned in the middle of the screen,
EWMH says that we must clip them to the monitor geometry. As such, this
will contain the logical screen workarea to the part in between those
two struts and leave off the rest of the screen. This breaks maximization
and placement in those cases.

(bzr r2844.5.1)