Rework the window shading system a bit.
There were a few changes in core to integrate better with window re-parenting which broke
the old method of doing shading. Now we update the shade state of the window when the window
is actually unmapped due to a shade and also ensure that the frame window is the right size
without touching the geometry of the client (such that CompWindow::geometry is representative
of the actual client geometry).

Also use the input extents last sent to server as a means to determine whether or not
window shading should be allowed rather than the extents last received as we don't update
the allowed window actions on input extents ConfigureNotify.

Fix LP#864478

Fixes LP: #864478

(bzr r2852.2.1)