Clean up and make the event tracking system into a proper object model and track events based on request serial numbers.
The request serial numbers at the moment are the only reliable way of matching
events with requests that we made, so use them for matching. This means that
the number of non-matches is going to go way down. At the moment, we still
look at whether or not we got an expected geometry and stack position back
from the server in the window's pending request queue and if we didn't then
some plugin is messing things up and we warn about that.

Fix LP #866752

Fixes LP: #866752

(bzr r2856.2.1)