Fixed excessive screen tearing still visible with "Sync To VBlank" enabled. (LP: #880707)
The composite plugin's frame timing logic did not correctly synchronize with
the vblank support of other plugins (opengl). This would lead to excessive
visible tearing even when "Sync To VBlank" is enabled in opengl.

This fix does NOT claim to eliminate all tearing, but it does eliminate almost
all of it. To completely eliminate tearing more work has to be done to
make the glXSwapBuffers code (opengl plugin) fast enough to use on every frame.

(bzr r2893.6.1)