Fix inaccurate frame timing causing tearing (LP: #880707)
Introduced "tickless" frame timing to the composite plugin. This means we
no longer need to poll for repaints. Composite only wakes up when a repaint
is required and it's the right time to do so. This not only improves vsync
accuracy, but also reduces CPU usage and power consumption.

(bzr r2893.4.1)