Revert revision 2938:     lp:~smspillaz/compiz-core/compiz-core.fix-880707-test-v3 in order to avoid a critical regression it causes (LP: #928655).
Unfortunately the regression was not just a logical bug, but a design bug.
The imposition of user-controlled frame rate throttling completely broke the
existing automatic frame rate throttling and detection. This was a design bug
because it attempted to separate waiting for Vsync from buffer swapping.
Doing so makes the detection of unthrottled rendering impossible. So the
paintscheduler would just switch to non-vsync mode and get stuck there
(LP: #928655).

Instead use the original algorithm from revision 2916:
    lp:~vanvugt/compiz-core/fix-880707.2
which is stable and has many months of community testing already.

(bzr r2984.3.1)