Fixed Super key release events sometimes not being sent to plugins (LP: #963470)
This regression was caused by the latest fix for LP: #806255, which has now
been removed. That was the third different fix attempted for #806255 and all
three have caused different regressions. So #806255 will not be fixed for
now. The only remaining fix (guaranteed to work) for that bug is to eliminate
and rewrite the XkbStateNotify code in compiz-core and several plugins.
However that is a massive change we can't afford to risk right now. So we
need to accept that LP: #806255 won't be fixed for a while.

Fixes LP: #963470

(bzr r3068.1.1)