Fix LP#969102 - Remove DecorWindow::computeShadowRegion and replace with a smarter class.
DecorClipGroup (split into GenericDecorClipGroup and MatchedDecorClipGroup) take a
DecorClippableInterface which where entry and exit from the group is guarded by the implementation
and are treated as one layer for the purposes of shadowing - eg the input region of each
member of the clip group will clip all of the shadows of the members, so that none of
the members appear to be overlapping.

This algorithm is a lot more optimized in that the shadows are only updated when
relevant windows are updated.

Tested too.

(bzr r3080.2.1)