Check outputRect when deciding whether to ignore custom clip regions.
It was possible for menus for repaint regions to never cover shadowRegion
which would mean that we'd never pass the shadow clip rects to
glDrawGeometry.

In reality, we only really care about not using shadowRegion in cases
where some plugin wants to paint us with an offset region which is actually
outside the window.

Fixes LP#978900

Fixes LP: #978900

(bzr r3087.1.1)