Don't insert the window into the server list above the window it was created above.
The server list might have been modified by the time that we process the
create event, and as such there is a case where a window can be inserted
into the server list above another window and not at the top (the default
for where windows are created) if that window is pending a restack.

When updateAttributes is called later, putting it above the correct window
will silently fail, because it is already there in the server list, even
though a restack was never issued to put it there.

(LP: #1008020) (LP: #886605)

Fixed bugs:
  - LP: #1008020
  - LP: #886605

(bzr r3237.5.1)