Avoid a potential race condition.
If any unhandled damage events that grow the bounding box remain in the queue
when we're rendering a frame, they will count as new damage we don't yet
know about (and so haven't redrawn). However calling XDamageSubtract with
repair==None tells the server we have repaired all damage when it's possible
we haven't repaired the full region described by those unhandled events.

The solution to this is to keep a copy of the bounding box of each damaged
drawable and only tell the server we've repaired the region that we actually
do know about and have repaired.

(bzr r3237.2.8)