Fixed CompTimerTestCallback.TimerOrder failure on slow systems (like valgrind) (LP: #1021139)
If the environment (like valgrind in ExperimentalMemCheck) is so slow that it
took more than 50ms to schedule the tests, then they would all be overdue
on startup (instead of just one of them) and the order would not be
predictable. So I spread the timers out over longer intervals and reduced
the number of repetitions to stop the new test from taking too long.

Fixes LP: #1021139

(bzr r3282.2.1)