Fix build failure due to invalid sed expression (LP: #1023738)
Don't use semicolon as the sed substitution delimiter:
    s;FIND;REPLACE;g
because cmake strips them and seems to convert them to spaces:
    s FIND REPLACE g

Usually this is safe, so long as FIND or REPLACE don't contain spaces. However
in the case of:
    s;FIND;;g
cmake will convert that to:
    s FIND g
which is an invalid sed expression.

Use colons instead, which seems safer and don't get stripped by cmake:
    s:FIND:REPLACE:g
    s:FIND::g

Fixes LP: #1023738

(bzr r3284.3.1)