Generate the CTestFile.cmake using cmake itself (LP: #1024214)
It was possible on a cmake run that enable_testing () would cause
CTestFile.cmake in a subdirectory to be overwritten and tests would
disappear from ctest as the file was not regenerated (as the rule
was not rebuilt). Force regeneration of this file every time cmake
is run.

(bzr r3285.1.1)