Don't execute events.
While we'll still get leaked callbacks, these are generally defended against
in the libraries and its better than bugs that happen due to re-entrancy.

I note here that switching between backends especially when callbacks are
queued is a really hard problem - while we do as much cleanup as possible
it seems as though its impossible to ensure that every callback will be removed.

(bzr r3279.4.433)