Handle negative values returned by viewportForGeometry correclty.
Previously if we got a negative value for viewportForGeometry, we'd try and
move to -n, -m, which would fail because checkDestination disallows those values.
      
checkDestination updated to accept negative values, because by default, we
actually pass negative values to it (since -dx, and -dy are negated before entry
into moveViewport). In any case, positive values are preferred, and the caller
was updated to provide positive values in case viewportForGeometry returns
negative values in order to avoid any unexpected behaviour.
      
In addition, if the window position needed to be changed, we'd be moving
the window by viewport-distance and not window distance due to a typo. That
was also fixed.

Fixes LP: #1037164

(bzr r3534.4.1)