Animation: Correctly specify the events an effect can be used with.
If AnimEffectUsedFor::none() is called, the effect will be marked to
be used with ALL events instead of NONE. This causes crashes if the
chosen random effect is not compatible with the triggering event.

For example dodge effect as random minimize animation causes segfault
as it's only supposed to be a focus in/out animation.

Fixed bugs:
  - LP: #1098185
  - LP: #1048840

(bzr r3554.3.1)