Added new test StackingSync.DestroyClientJustBeforeReparent
Test that compiz' internal stack representation does not become
confused when a a client window is marked as destroyed just before
the ReparentNotify arrives. The parent window should not be added to
the window stack twice, and instead by added only once when we
recieve a CreateNotify for it.

As such, restacking the parent window toward the bottom of the stack
should result in any dependent stacking operations suceeding and not being
mistakenly placed at the top of the stack.

(bzr r3745.2.1)