Window: compute the correct output using the absolute window position
Window geometry has absolute coordinates that include viewports, but we
need to use viewport relative coordinates when figuring out the window output
or the window will be considered offscreen and thus it will fallback to first monitor.
This is generally applied, but was not done well in addWindowSizeChanges which
was causing Fullscreen and (horizontally) maximized windows to be moved to the
primary monitor when activated.

Fixes LP: #1475721

(bzr r3958.6.2)