opengl: Don't invalidate the scratch buffer frame age every time it is queried
We only want to do this when we query the scratch buffer age and then
use it - as it is only on use when the age resets back to an invalidated
state.

Previously, the age would re-set every time it is queried, which would
cause it to appear invalidated the next time damageCutoff was reached.

(bzr r4010.6.10)