animationaddon: Port the animationaddon plugin to use modern GL API.
This involved a lot of changes:
- Removal of glPush/glPop. State is assumed to be off in newer versions
of compiz, so switch off whatever gets switched on.
- Switch from glVertexPointer/glDrawArrays/glDrawElements to GLVertexBuffer
- Changes in primitive assembly: GL_POLYGON doesn't exist in GLES
so switch to using GL_TRIANGLES. This meant that the indices
for all the animatons needed to be re-tesselated. That was done
by hand using a winding rule which just duplicates the first and
third vertex around the fourth.
- Dropping of glTexEnv calls.
- Reworking of the depth test, since writes to the depth buffer
have been disabled by default.

(bzr r4061)