Resize: add options to use averageColor for dawing shapes
If the average color for the desktop is defined, let's use it
for painting the alternative view of move window.
Added also an option (on by default) to draw a larger border
using a lighter or darker color (depending on the current color
average), to make all this more visible.

(bzr r4121.2.16)