Do not enable "Target Focused Window on Specific Zoom" by default
It is much more useful to have the option "Target Focused Window on Specific
Zoom" disabled by default: specific zoom are typically used to quickly switch
between a global view and a view on a specific part of the screen. It is
more convenient for the user to use the mouse to point which part of the
screen should be magnified, instead of the focused window which may be much
bigger/smaller/shifted/etc. compared to what should be magnified.

Fixes LP: #1763692

(bzr r4143.10.1)